% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgd.ds.c}
\alias{lgd.ds.c}
\title{Synthetic modeling dataset}
\format{
An object of class \code{data.frame} with 1200 rows and 20 columns.
}
\usage{
lgd.ds.c
}
\description{
Synthetic modeling dataset of observed LGD values for contracts with complete recovery process.
Dataset consists of 1200 observations and 19 risk factors.
}
\keyword{datasets}
